1.Fix the build error caused by the modification for AC6 2.Revise startup_M051.s and retarget.c files to fix build issue. 3.Modify settings of Keil project to speed up compilation. 4.Revise copyright format for gen. chm file.

+++ SampleCode/StdDriver/ADC_PwmTrigger/main.c
@@ -6,8 +6,9 @@
  * @brief    Demonstrate how to trigger ADC by PWM.
  *
  * @note
- * Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
+ * @copyright SPDX-License-Identifier: Apache-2.0
  *
+ * @copyright Copyright (C) 2014 Nuvoton Technology Corp. All rights reserved.
  ******************************************************************************/
 #include <stdio.h>
 #include "M051Series.h"
@@ -159,7 +160,7 @@ void ADC_PWMTrigTest_SingleOpMode()
     /* Clear the ADC interrupt flag */
     ADC_CLR_INT_FLAG(ADC, ADC_ADF_INT);
 
-    printf("Channel 2: 0x%X\n", ADC_GET_CONVERSION_DATA(ADC, 2));
+    printf("Channel 2: 0x%X\n",(unsigned int) ADC_GET_CONVERSION_DATA(ADC, 2));
 
     /* Disable ADC */
     ADC_POWER_DOWN(ADC);
